feat(checkin): Add check-in button to landing page
Adds the front-end logic to display an event check-in button in the up coming
events section if the event matches the check-in day and a check-in link was
added to the brigade. The button will replace the event and be displayed on the
day of the event and will disappear when the event is over.