fix(projects): added active flag, various fixes
this adds the much needed active flag functionality, along with the
ability to update the active tag in edit, create and manage, as well
as the ability to filter according to in on projects list. it also adds
a big caps "INACTIVE PROJECT" to the sidebar of project pages if not
flagged. This also adds the flag to the projects model and the
corresponding seed file.

Additionally, this fixes the long-standing problem of the create project
page being wildly outdated. Additionally, this is fixing the very bad
bug of the server responding with success before the save is complete.
Additionally this fixes a few minor styling bugs on the project list
page.

Theres also a current issue with manage projects page not saving data
correctly, that will need to be addressed later.

Screenshots:

![](https://i.imgur.com/0IPskUD.jpg)
![](https://i.imgur.com/GFPsHrY.png)
![](https://i.imgur.com/kwcvMm8.png)
![](https://i.imgur.com/uwBzKTL.jpg)
![](https://i.imgur.com/hO54G72.png)
![](https://i.imgur.com/uAQw67n.png)
![](https://i.imgur.com/RycGa5S.png)

cc @sfbrigade/brigadehub